requirements.txt
attrs==24.2.0
Automat==22.10.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.15.4
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jmespath==1.0.1
lxml==5.2.2
packaging==24.1
parsel==1.9.1
Protego==0.3.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.2.1
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
service-identity==24.1.0
six==1.16.0
SQLAlchemy==2.0.32
tldextract==5.1.2
tomli==2.0.1
Twisted==24.7.0
typing_extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
zope.interface==7.0.1
